// client/stocklinx app/src/redux/component/constant.ts
export enum componentConst {
  FETCH_COMPONENTS_REQUEST = "FETCH_COMPONENTS_REQUEST",
  FETCH_COMPONENTS_SUCCESS = "FETCH_COMPONENTS_SUCCESS",
  FETCH_COMPONENTS_FAILURE = "FETCH_COMPONENTS_FAILURE",
  FETCH_COMPONENT_REQUEST = "FETCH_COMPONENT_REQUEST",
  FETCH_COMPONENT_SUCCESS = "FETCH_COMPONENT_SUCCESS",
  FETCH_COMPONENT_FAILURE = "FETCH_COMPONENT_FAILURE",
  CREATE_COMPONENT_REQUEST = "CREATE_COMPONENT_REQUEST",
  CREATE_COMPONENT_SUCCESS = "CREATE_COMPONENT_SUCCESS",
  CREATE_COMPONENT_FAILURE = "CREATE_COMPONENT_FAILURE",
  UPDATE_COMPONENT_REQUEST = "UPDATE_COMPONENT_REQUEST",
  UPDATE_COMPONENT_SUCCESS = "UPDATE_COMPONENT_SUCCESS",
  UPDATE_COMPONENT_FAILURE = "UPDATE_COMPONENT_FAILURE",
  REMOVE_COMPONENT_REQUEST = "REMOVE_COMPONENT_REQUEST",
  REMOVE_COMPONENT_SUCCESS = "REMOVE_COMPONENT_SUCCESS",
  REMOVE_COMPONENT_FAILURE = "REMOVE_COMPONENT_FAILURE",
  SET_COMPONENT = "SET_COMPONENT",
  CLEAR_COMPONENT = "CLEAR_COMPONENT",
  SET_COMPONENTS = "SET_COMPONENTS",
  CLEAR_COMPONENTS = "CLEAR_COMPONENTS",
}